+++ src/store/loading/reducers.js
@@ -1,9 +1,10 @@
 import { handleActions } from 'redux-actions';
 import { requestAddressList, requestAddressListFailure, requestAddressListSuccess } from '../addressList';
 import { authUserFailure, authUserRequest, authUserSuccess } from '../auth';
+import { logoutSuccess } from '../logout';
 import { updateProfileFailure, updateProfile, updateProfileSuccess, requestProfile, requestProfileSuccess, requestProfileFailure } from '../profile';
 import { registerUserFailure, registerUserRequest, registerUserSuccess } from '../register';
-import { requestRoute, requestRouteFailure, requestRouteSuccess, cancelRequestRoute } from '../route/actions';
+import { requestRoute, requestRouteFailure, requestRouteSuccess, cancelRequestRoute } from '../route';
 
 export const isLoadingReducer = handleActions({
     [authUserRequest]: () => true,
@@ -25,4 +26,5 @@ export const isLoadingReducer = handleActions({
     [requestRouteSuccess]: () => false,
     [requestRouteFailure]: () => false,
     [cancelRequestRoute]: () => false,
+    [logoutSuccess]: () => false,
 }, { isLoading: false });
